requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
python-jose==3.3.0
jupyter==1.0.0
notebook==7.1.0
Flask-Limiter==3.5.0
cachetools==5.3.2
Werkzeug==3.0.1 